Avoid unnecessary conversions
No need to convert values here.

+++ assert/assertions.go
@@ -1014,7 +1014,7 @@ func toFloat(x interface{}) (float64, bool) {
 	case float32:
 		xf = float64(xn)
 	case float64:
-		xf = float64(xn)
+		xf = xn
 	case time.Duration:
 		xf = float64(xn)
 	default:
